Address some code review comments
- Update CHANGELOG
- Small refactoring of SettingsHelper
- Remove redundant  prefix from  protocol
- Move XCTestCase+Extras.swift to TuistCoreTesting
